Added Authorization

+++ backend/routes/posts.js
@@ -82,8 +82,11 @@ router.get('', (req, res, next) => {
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId  // got from checkAuth middleware
     });
+    //console.log(req.userData);
+    //return res.status(200).json({});
     post.save()
     .then(createdPost => {
       res.status(201).json({
@@ -107,20 +110,31 @@ router.get('', (req, res, next) => {
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id}, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId}, post)
     .then(result => {
-      res.status(200).json({message: 'Update Successful'});
+      if (result.nModified > 0) {
+        res.status(200).json({message: 'Update Successful'});
+      } else {
+        res.status(401).json({message: 'Not Authorized'});
+      }
+      
     });
   });
   
   router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id})
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId})
     .then(result => {
       console.log(result);
-      res.status(200).json({message: 'Post deleted!'});
+      if (result.n > 0) {
+        res.status(200).json({message: 'Post deleted!'});
+      } else {
+        res.status(401).json({message: 'Not authorized'});
+      }
+      
     });
   });
